feat: 동시성 예외 처리 추가 및 최근 검색어 API 엔드포인트 분리 (#18)
